+++ src/mtx_generation/rndgen.h
@@ -1,8 +1,8 @@
 #ifndef __MTX__
 #define __MTX__
 
-void generate_real_value_matrix(float *matrix,int rows, int cols);
-void generate_zero_value_matrix(float *matrix,int rows, int cols);
+void generate_real_value_matrix(double *matrix,int rows, int cols);
+void generate_zero_value_matrix(double *matrix,int rows, int cols);
 
 
 #endif
